# Propose going back to condition variable to notify writeLoop